+++ R/investr.R
@@ -6,22 +6,23 @@
 #' response) to make inference on the corresponding unknown value of the 
 #' explanatory variable. 
 #'
-#' A detailed \href{https://journal.r-project.org/archive/2014/RJ-2014-009/index.html}{introduction to investr} has been published in The R Journal: 
+#' A detailed [introduction to investr](https://journal.r-project.org/archive/2014/RJ-2014-009/index.html) has been published in The R Journal: 
 #' "investr: An R Package for Inverse Estimation." You 
-#' can track development at \url{https://github.com/bgreenwell/investr}. To report 
+#' can track development at <https://github.com/bgreenwell/investr>. To report 
 #' bugs or issues, contact the main author directly or submit them to 
-#' \url{https://github.com/bgreenwell/investr/issues}. 
+#' <https://github.com/bgreenwell/investr/issues>. 
 #'
-#' As of right now, \code{investr} supports (univariate) inverse estimation 
+#' As of right now, `investr` supports (univariate) inverse estimation 
 #' with objects of class:
 #' \itemize{
-#'   \item{\code{lm}} --- linear models (multiple predictor variables allowed)
-#'   \item{\code{glm}} --- generalized linear models (multiple predictor variables allowed)
-#'   \item{\code{nls}} --- nonlinear least-squares models
-#'   \item{\code{lme}} --- linear mixed-effects models (fit using the 
-#'     \code{nlme} package)
+#'   \item{`lm`} --- linear models (multiple predictor variables allowed)
+#'   \item{`glm`} --- generalized linear models (multiple predictor variables allowed)
+#'   \item{`nls`} --- nonlinear least-squares models
+#'   \item{`lme`} --- linear mixed-effects models (fit using the
+#'     `nlme` package)
+#'   \item{`survfit`/`Surv`} --- Kaplan-Meier survival curves (fit using the
+#'     `survival` package); see [invest.survfit()]
 #' }
 #' 
-#' @docType package
-#' @name investr
-NULL
+#' @keywords internal
+"_PACKAGE"