+++ inst/tinytest/test_utilities.R
@@ -1,0 +1,3 @@
+# Utility functions
+
+expect_identical(investr:::makeData(1:3, label = "a"), data.frame("a" = 1:3))